# apps/api/src/main/java/com/kaas/api/controlplane/domain/SnapshotFeature.java
package com.kaas.api.controlplane.domain;

import java.util.UUID;

public record SnapshotFeature(
        UUID featureId,
        UUID revisionId,
        long revisionNumber,
        String logicalPath,
        String sourceDigest) {}