# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: KaaS Control Plane API
  version: 1.0.0
  description: >-
    Project, immutable FeatureRevision, SecretReference metadata, immutable EnvironmentRevision, and immutable
    RunProfileRevision operations and TestRun create/get/list/snapshot are implemented. Run cancellation, events,
    results, and artifacts remain proposed contracts only. Authentication and tenant authorization are required;
    arbitrary execution remains disabled. Implemented JSON mutation requests are
    limited to 1048576 encoded bytes, and decoded Feature source is limited to 524288 UTF-8 bytes.
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/projects/{projectId}/runs:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Runs]
      operationId: createRun
      summary: Persist a CREATED test-run intent and immutable snapshot
      x-kaas-implementation-status: IMPLEMENTED
      description: >-
        Atomically records a CREATED run, its immutable execution snapshot, and the scoped idempotency record. The
        response does not imply that a command was generated or published, or that the run reached QUEUED. An
        identical replay returns the original status, body, Location, and ETag. Reusing the
        same key in the same authenticated organization, operation, and project scope with a different normalized
        request fingerprint returns an idempotency conflict.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '202':
          description: Run intent was durably accepted; scheduling may not have started.
          headers:
            Location:
              $ref: '#/components/headers/RunLocation'
            ETag:
              $ref: '#/components/headers/RunEtag'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationProblem'
    get:
      tags: [Runs]
      operationId: listProjectRuns
      summary: List test runs in one authorized project
      x-kaas-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Stable page ordered by createdAt descending and runId descending.
          headers:
            Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationProblem'}
  /api/v1/runs/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      operationId: getRun
      summary: Get the canonical run representation
      x-kaas-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Current run state and independent outcome dimensions.
          headers:
            ETag:
              $ref: '#/components/headers/RunEtag'
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/runs/{runId}/snapshot:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      operationId: getRunSnapshot
      summary: Get the immutable materialized execution snapshot
      x-kaas-implementation-status: IMPLEMENTED
      description: >-
        Returns exact revision provenance, typed effective non-secret configuration, metadata-only secret references,
        profile settings, and the server-owned engine descriptor. It never resolves secret values or creates an
        execution command.
      responses:
        '200':
          description: The persisted snapshot; the read path does not re-resolve mutable inputs.
          headers:
            ETag:
              description: Strong tag derived from the immutable snapshot digest.
              schema: {type: string}
            Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunSnapshot'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API-client authentication; the token is never accepted in a query string.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque client key scoped to authenticated organization, principal, operation, and target resource path. Implemented creation keys currently have no expiry.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._~-]+$'
...
  headers:
...
    IdempotencyReplayed:
      description: Whether this response replays an earlier committed success.
      schema: {type: string, enum: ['true', 'false']}
    CacheControlNoStore:
      description: Tenant-sensitive configuration metadata must not be stored by shared or private caches.
      schema: {type: string, const: no-store}
    RunLocation:
      description: Canonical run resource.
      schema:
        type: string
        pattern: '^/api/v1/runs/[0-9a-fA-F-]{36}$'
    RunEtag:
      description: Strong representation tag derived from runVersion, for example "run-3".
      schema:
        type: string
...
  responses:
    BadRequest:
      description: JSON syntax, duplicate keys, or request encoding is invalid.
      headers:
        Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ProblemDetail'}
    Unauthorized:
      description: Authentication is missing or invalid.
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema: {type: string, const: Bearer}
        Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotFound:
      description: The authorized resource does not exist, or existence is intentionally concealed.
      headers:
        Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    IdempotencyConflict:
      description: The scoped key was already used with a different normalized request fingerprint.
      headers:
        Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    ValidationProblem:
      description: Request validation failed. Submitted values are never echoed.
      headers:
        Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    PayloadTooLarge:
      description: The encoded or decoded request exceeds the contractual size limit.
      headers:
        Cache-Control: {$ref: '#/components/headers/CacheControlNoStore'}
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ConfigurationVariable:
      oneOf:
        - $ref: '#/components/schemas/StringConfigurationVariable'
        - $ref: '#/components/schemas/IntegerConfigurationVariable'
        - $ref: '#/components/schemas/BooleanConfigurationVariable'
      discriminator:
        propertyName: type
        mapping:
          STRING: '#/components/schemas/StringConfigurationVariable'
          INTEGER: '#/components/schemas/IntegerConfigurationVariable'
          BOOLEAN: '#/components/schemas/BooleanConfigurationVariable'
    StringConfigurationVariable:
      type: object
      required: [key, type, value]
      properties:
        key: {$ref: '#/components/schemas/ConfigurationKey'}
        type: {const: STRING}
        value:
          type: string
          description: Exact control-free, well-formed Unicode; at most 4096 UTF-8 bytes with no normalization.
      additionalProperties: false
    IntegerConfigurationVariable:
      type: object
      required: [key, type, value]
      properties:
        key: {$ref: '#/components/schemas/ConfigurationKey'}
        type: {const: INTEGER}
        value:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
      additionalProperties: false
    BooleanConfigurationVariable:
      type: object
      required: [key, type, value]
      properties:
        key: {$ref: '#/components/schemas/ConfigurationKey'}
        type: {const: BOOLEAN}
        value: {type: boolean}
      additionalProperties: false
    ConfigurationKey:
      type: string
      pattern: '^[A-Za-z_][A-Za-z0-9_.-]{0,127}$'
      description: Exact case-sensitive portable key; duplicate plain/secret keys in one revision are rejected.
    SecretBinding:
      type: object
      description: Binds a key to non-capability SecretReference metadata. It contains no secret value.
      required: [key, secretReferenceId]
      properties:
        key: {$ref: '#/components/schemas/ConfigurationKey'}
        secretReferenceId: {type: string, format: uuid}
      additionalProperties: false
...
    ProfileSelection:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          maxItems: 100
          uniqueItems: true
          items: {type: string, pattern: '^@[A-Za-z0-9_.:-]{1,127}$'}
      additionalProperties: false
...
    Sha256Digest:
      type: string
      pattern: '^sha256:[a-f0-9]{64}$'
      description: Server-computed digest of the documented versioned canonical content representation.
    PageMetadata:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 100}
        totalElements: {type: integer, minimum: 0}
        totalPages: {type: integer, minimum: 0}
      additionalProperties: true
    CreateRunRequest:
      type: object
      description: >-
        IMPLEMENTED profile-based run intent. The immutable RunProfileRevision supplies its pinned EnvironmentRevision
        and all execution settings. Client order is non-semantic; duplicate revision IDs are rejected.
      required: [featureRevisionIds, runProfileRevisionId]
      properties:
        featureRevisionIds:
          type: array
          minItems: 1
          maxItems: 1000
          uniqueItems: true
          items:
            type: string
            format: uuid
        runProfileRevisionId:
          type: string
          format: uuid
      additionalProperties: false
...
    ScenarioRetry:
      type: object
      required: [maxAttempts, delayMilliseconds]
      properties:
        maxAttempts:
          type: integer
          minimum: 1
          maximum: 5
        delayMilliseconds:
          type: integer
          minimum: 0
          maximum: 30000
      additionalProperties: false
      description: Karate scenario retry inside one execution attempt; never an orchestration retry.
    ArtifactPolicy:
      type: object
      required: [types, maxArtifactBytes, maxTotalBytes]
      properties:
        types:
          type: array
          maxItems: 4
          uniqueItems: true
          items:
            enum: [KARATE_HTML_REPORT, RAW_RESULT, EXECUTION_LOG, OTHER]
        maxArtifactBytes:
          type: integer
          minimum: 0
          maximum: 104857600
        maxTotalBytes:
          type: integer
          minimum: 0
          maximum: 524288000
      additionalProperties: false
    Run:
      type: object
      description: >-
        Current public TestRun aggregate. The implemented create slice always returns version 1 in CREATED with no
        attempt or queue deadline; queue timing starts only when a future scheduler performs CREATED to QUEUED.
      required: [runId, projectId, runVersion, lifecycleState, cancellationStatus, testOutcome, infrastructureOutcome, qualityGateStatus, snapshotDigest, createdBy, createdAt, updatedAt]
      properties:
        runId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        runVersion:
          type: integer
          minimum: 1
        lifecycleState:
          enum: [CREATED, QUEUED, CLAIMED, PROVISIONING, RUNNING, COLLECTING_RESULTS, PROCESSING_RESULTS, STOPPING, COMPLETED]
        cancellationStatus:
          enum: [NOT_REQUESTED, REQUESTED, ACKNOWLEDGED]
        testOutcome:
          type: [string, 'null']
          enum: [PASSED, FAILED, NOT_AVAILABLE, null]
        infrastructureOutcome:
          type: [string, 'null']
          enum: [SUCCEEDED, FAILED, TIMED_OUT, CANCELLED, null]
        qualityGateStatus:
          enum: [NOT_EVALUATED, PASSED, FAILED]
        snapshotDigest:
          $ref: '#/components/schemas/Sha256Digest'
        createdBy:
          type: string
          minLength: 1
          maxLength: 255
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    RunPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/Run'}
    RunSnapshot:
      type: object
      required: [runId, projectId, snapshotVersion, features, environment, runProfile, effectiveConfiguration, secretBindings, selection, parallelism, scenarioRetry, executionTimeoutSeconds, artifactPolicy, engine, snapshotDigest]
      properties:
        runId: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        snapshotVersion: {type: integer, const: 1}
        features:
          type: array
          minItems: 1
          maxItems: 1000
          items: {$ref: '#/components/schemas/SnapshotFeature'}
        environment: {$ref: '#/components/schemas/SnapshotRevision'}
        runProfile: {$ref: '#/components/schemas/SnapshotRevision'}
        effectiveConfiguration:
          type: array
          maxItems: 500
          items: {$ref: '#/components/schemas/ConfigurationVariable'}
        secretBindings:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/SecretBinding'}
        selection: {$ref: '#/components/schemas/ProfileSelection'}
        parallelism: {type: integer, minimum: 1, maximum: 32}
        scenarioRetry: {$ref: '#/components/schemas/ScenarioRetry'}
        executionTimeoutSeconds: {type: integer, minimum: 1, maximum: 3600}
        artifactPolicy: {$ref: '#/components/schemas/ArtifactPolicy'}
        engine: {$ref: '#/components/schemas/EngineDescriptor'}
        snapshotDigest: {$ref: '#/components/schemas/Sha256Digest'}
      additionalProperties: false
    SnapshotFeature:
      type: object
      required: [featureId, revisionId, revisionNumber, logicalPath, sourceDigest]
      properties:
        featureId: {type: string, format: uuid}
        revisionId: {type: string, format: uuid}
        revisionNumber: {type: integer, minimum: 1}
        logicalPath: {type: string, minLength: 1, maxLength: 512}
        sourceDigest: {$ref: '#/components/schemas/Sha256Digest'}
      additionalProperties: false
    SnapshotRevision:
      type: object
      required: [resourceId, revisionId, revisionNumber, contentDigest]
      properties:
        resourceId: {type: string, format: uuid}
        revisionId: {type: string, format: uuid}
        revisionNumber: {type: integer, minimum: 1}
        contentDigest: {$ref: '#/components/schemas/Sha256Digest'}
      additionalProperties: false
    EngineDescriptor:
      type: object
      required: [engine, version]
      properties:
        engine: {type: string, const: KARATE}
        version:
          type: string
          pattern: '^[0-9]+\.[0-9]+\.[0-9]+(?:[-+][0-9A-Za-z.-]+)?$'
      additionalProperties: false
...
    ProblemDetail:
      type: object
      required: [type, title, status, code, detail, instance, requestId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          maxLength: 128
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          enum: [UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, VALIDATION_FAILED, IDEMPOTENCY_CONFLICT, PROJECT_NAME_CONFLICT, FEATURE_PATH_CONFLICT, SECRET_REFERENCE_NAME_CONFLICT, ENVIRONMENT_NAME_CONFLICT, RUN_PROFILE_NAME_CONFLICT, CONFLICT, LIFECYCLE_CONFLICT, RESULT_NOT_READY, UNSUPPORTED_OPERATION, PAYLOAD_TOO_LARGE, RATE_LIMITED, EVENT_HISTORY_GAP, INTERNAL_ERROR]
        detail:
          type: string
          maxLength: 2048
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          minLength: 1
          maxLength: 128
      additionalProperties: true
      description: RFC 9457-style problem details. Extensions must never expose submitted values, secrets, stack traces, host paths, or concealed tenant identity.
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemDetail'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              maxItems: 100
              items:
                type: object
                required: [pointer, message]
                properties:
                  pointer:
                    type: string
                    pattern: '^(/([^/~]|~[01])*)*$'
                    maxLength: 512
                  message:
                    type: string
                    maxLength: 512
                additionalProperties: false
...